+++ mobile/android/android-components/components/feature/session-bundling/src/main/java/mozilla/components/feature/session/bundling/SessionBundleStorage.kt
@@ -14,11 +14,15 @@ import android.support.annotation.VisibleForTesting
 import android.support.annotation.WorkerThread
 import mozilla.components.browser.session.Session
 import mozilla.components.browser.session.SessionManager
+import mozilla.components.browser.session.ext.writeSnapshot
 import mozilla.components.browser.session.storage.AutoSave
+import mozilla.components.concept.engine.Engine
 import mozilla.components.feature.session.bundling.adapter.SessionBundleAdapter
 import mozilla.components.feature.session.bundling.db.BundleDatabase
 import mozilla.components.feature.session.bundling.db.BundleEntity
 import mozilla.components.feature.session.bundling.ext.toBundleEntity
+import mozilla.components.support.ktx.java.io.truncateDirectory
+import java.io.File
 import java.lang.IllegalArgumentException
 import java.util.concurrent.TimeUnit
 
@@ -30,7 +34,8 @@ import java.util.concurrent.TimeUnit
  */
 @Suppress("TooManyFunctions")
 class SessionBundleStorage(
-    context: Context,
+    private val context: Context,
+    private val engine: Engine,
     internal val bundleLifetime: Pair<Long, TimeUnit>
 ) : AutoSave.Storage {
     @VisibleForTesting(otherwise = VisibleForTesting.PRIVATE)
@@ -54,7 +59,7 @@ class SessionBundleStorage(
             .getLastBundle(since)
             .also { lastBundle = it }
 
-        return entity?.let { SessionBundleAdapter(it) }
+        return entity?.let { SessionBundleAdapter(context, engine, it) }
     }
 
     /**
@@ -91,6 +96,8 @@ class SessionBundleStorage(
         }
 
         bundle.actual.let { database.bundleDao().deleteBundle(it) }
+
+        bundle.actual.stateFile(context, engine).delete()
     }
 
     /**
@@ -101,6 +108,9 @@ class SessionBundleStorage(
     fun removeAll() {
         lastBundle = null
         database.clearAllTables()
+
+        getStateDirectory(context)
+            .truncateDirectory()
     }
 
     /**
@@ -109,7 +119,7 @@ class SessionBundleStorage(
      */
     @Synchronized
     fun current(): SessionBundle? {
-        return lastBundle?.let { SessionBundleAdapter(it) }
+        return lastBundle?.let { SessionBundleAdapter(context, engine, it) }
     }
 
     /**
@@ -146,7 +156,7 @@ class SessionBundleStorage(
             .bundleDao()
             .getBundles(since, limit)
         ) { list ->
-            list.map { SessionBundleAdapter(it) }
+            list.map { SessionBundleAdapter(context, engine, it) }
         }
     }
 
@@ -166,7 +176,7 @@ class SessionBundleStorage(
         return database
             .bundleDao()
             .getBundlesPaged(since)
-            .map { entity -> SessionBundleAdapter(entity) }
+            .map { entity -> SessionBundleAdapter(context, engine, entity) }
     }
 
     /**
@@ -203,7 +213,12 @@ class SessionBundleStorage(
             return
         }
 
-        val bundle = snapshot.toBundleEntity().also { lastBundle = it }
+        val bundle = snapshot.toBundleEntity().also {
+            lastBundle = it
+        }
+
+        bundle.stateFile(context, engine).writeSnapshot(snapshot)
+
         bundle.id = database.bundleDao().insertBundle(bundle)
     }
 
@@ -214,8 +229,10 @@ class SessionBundleStorage(
         if (snapshot.isEmpty()) {
             // If this snapshot is empty then instead of saving an empty bundle: Remove the bundle. Otherwise
             // we end up with empty bundles to restore and that is not helpful at all.
-            remove(SessionBundleAdapter(bundle))
+            remove(SessionBundleAdapter(context, engine, bundle))
         } else {
+            bundle.stateFile(context, engine).writeSnapshot(snapshot)
+
             bundle.updateFrom(snapshot)
             database.bundleDao().updateBundle(bundle)
         }
@@ -223,4 +240,12 @@ class SessionBundleStorage(
 
     @VisibleForTesting(otherwise = VisibleForTesting.PRIVATE)
     internal fun now() = System.currentTimeMillis()
+
+    companion object {
+        internal fun getStateDirectory(context: Context): File {
+            return File(context.filesDir, "mozac.feature.session.bundling").apply {
+                mkdirs()
+            }
+        }
+    }
 }