+++ mobile/android/android-components/components/feature/session-bundling/src/main/java/mozilla/components/feature/session/bundling/adapter/SessionBundleAdapter.kt
@@ -4,21 +4,22 @@
 
 package mozilla.components.feature.session.bundling.adapter
 
+import android.content.Context
 import mozilla.components.browser.session.SessionManager
-import mozilla.components.browser.session.storage.SnapshotSerializer
+import mozilla.components.browser.session.ext.readSnapshot
 import mozilla.components.concept.engine.Engine
 import mozilla.components.feature.session.bundling.SessionBundle
 import mozilla.components.feature.session.bundling.db.BundleEntity
-import org.json.JSONException
 
 /**
  * Adapter implementation to make a [BundleEntity] accessible as [SessionBundle] without "leaking" the underlying
  * entity class.
  */
 internal class SessionBundleAdapter(
+    internal val context: Context,
+    private val engine: Engine,
     internal val actual: BundleEntity
 ) : SessionBundle {
-
     override val id: Long?
         get() = actual.id
 
@@ -31,11 +32,8 @@ internal class SessionBundleAdapter(
     /**
      * Re-create the [SessionManager.Snapshot] from the state saved in the database.
      */
-    override fun restoreSnapshot(engine: Engine): SessionManager.Snapshot? {
-        return try {
-            SnapshotSerializer().fromJSON(engine, actual.state)
-        } catch (e: JSONException) {
-            null
-        }
+    override fun restoreSnapshot(): SessionManager.Snapshot? {
+        return actual.stateFile(context, engine)
+            .readSnapshot(engine)
     }
 }