+++ mobile/android/android-components/components/feature/session-bundling/src/main/java/mozilla/components/feature/session/bundling/db/BundleEntity.kt
@@ -7,9 +7,17 @@ package mozilla.components.feature.session.bundling.db
 import android.arch.persistence.room.ColumnInfo
 import android.arch.persistence.room.Entity
 import android.arch.persistence.room.PrimaryKey
+import android.content.Context
+import android.support.annotation.VisibleForTesting
+import android.support.annotation.VisibleForTesting.PRIVATE
+import android.support.annotation.WorkerThread
+import android.util.AtomicFile
 import mozilla.components.browser.session.SessionManager
-import mozilla.components.browser.session.storage.SnapshotSerializer
+import mozilla.components.concept.engine.Engine
 import mozilla.components.feature.session.bundling.SessionBundle
+import mozilla.components.feature.session.bundling.SessionBundleStorage
+import java.io.File
+import java.util.UUID
 
 /**
  * Internal entity representing a session bundle as it gets saved to the database. This class implements [SessionBundle]
@@ -21,24 +29,32 @@ internal data class BundleEntity(
     @ColumnInfo(name = "id")
     var id: Long?,
 
-    @ColumnInfo(name = "state")
-    var state: String,
-
     @ColumnInfo(name = "saved_at")
     var savedAt: Long,
 
     @ColumnInfo(name = "urls")
-    var urls: UrlList
+    var urls: UrlList,
+
+    @ColumnInfo(name = "file")
+    val file: String = UUID.randomUUID().toString()
 ) {
     /**
      * Updates this entity with the value from the given snapshot.
      */
     fun updateFrom(snapshot: SessionManager.Snapshot): BundleEntity {
-        state = SnapshotSerializer().toJSON(snapshot)
         savedAt = System.currentTimeMillis()
         urls = UrlList(snapshot.sessions.map { it.session.url })
         return this
     }
+
+    @WorkerThread
+    fun stateFile(context: Context, engine: Engine): AtomicFile {
+        return AtomicFile(statePath(context, engine))
+    }
+
+    @VisibleForTesting(otherwise = PRIVATE)
+    internal fun statePath(context: Context, engine: Engine) =
+        File(SessionBundleStorage.getStateDirectory(context), engine.name() + '_' + file)
 }
 
 internal data class UrlList(