+++ mobile/android/android-components/components/feature/session-bundling/src/test/java/mozilla/components/feature/session/bundling/adapter/SessionBundleAdapterTest.kt
@@ -4,8 +4,11 @@
 
 package mozilla.components.feature.session.bundling.adapter
 
+import android.content.Context
+import androidx.test.core.app.ApplicationProvider
 import mozilla.components.browser.session.Session
 import mozilla.components.browser.session.SessionManager
+import mozilla.components.browser.session.ext.writeSnapshot
 import mozilla.components.feature.session.bundling.db.BundleEntity
 import mozilla.components.feature.session.bundling.db.UrlList
 import mozilla.components.support.test.mock
@@ -18,17 +21,21 @@ import org.robolectric.RobolectricTestRunner
 
 @RunWith(RobolectricTestRunner::class)
 class SessionBundleAdapterTest {
+    private val context: Context
+        get() = ApplicationProvider.getApplicationContext()
+
     @Test
     fun `restoreSnapshot restores snapshot from state`() {
-        val bundle = BundleEntity(0, "", 0, UrlList(listOf()))
+        val bundle = BundleEntity(0, 0, UrlList(listOf()))
 
         val snapshot = SessionManager.Snapshot(
             listOf(SessionManager.Snapshot.Item(session = Session("https://www.mozilla.org"))),
             selectedSessionIndex = 0)
 
         bundle.updateFrom(snapshot)
+        bundle.stateFile(context, mock()).writeSnapshot(snapshot)
 
-        val restoredSnapshot = SessionBundleAdapter(bundle).restoreSnapshot(mock())
+        val restoredSnapshot = SessionBundleAdapter(context, mock(), bundle).restoreSnapshot()
 
         assertNotNull(restoredSnapshot!!)
 
@@ -39,20 +46,20 @@ class SessionBundleAdapterTest {
 
     @Test
     fun `Accessing id through adapter`() {
-        val bundle = BundleEntity(42, "", 0, UrlList(listOf()))
-        val adapter = SessionBundleAdapter(bundle)
+        val bundle = BundleEntity(42, 0, UrlList(listOf()))
+        val adapter = SessionBundleAdapter(context, mock(), bundle)
 
         assertEquals(42L, adapter.id)
     }
 
     @Test
     fun `Accessing list of URLs through adapter`() {
-        val bundle = BundleEntity(42, "", 0, UrlList(listOf(
+        val bundle = BundleEntity(42, 0, UrlList(listOf(
             "https://www.mozilla.org",
             "https://www.firefox.com"
         )))
 
-        val adapter = SessionBundleAdapter(bundle)
+        val adapter = SessionBundleAdapter(context, mock(), bundle)
 
         assertEquals(2, adapter.urls.size)
         assertEquals("https://www.mozilla.org", adapter.urls[0])
@@ -61,9 +68,9 @@ class SessionBundleAdapterTest {
 
     @Test
     fun `Accessing last save date through adapter`() {
-        val bundle = BundleEntity(42, "", 1548165508982, UrlList(listOf()))
+        val bundle = BundleEntity(42, 1548165508982, UrlList(listOf()))
 
-        val adapter = SessionBundleAdapter(bundle)
+        val adapter = SessionBundleAdapter(context, mock(), bundle)
 
         assertEquals(1548165508982, adapter.lastSavedAt)
     }