+++ mobile/android/android-components/components/feature/session-bundling/src/test/java/mozilla/components/feature/session/bundling/db/BundleEntityTest.kt
@@ -6,7 +6,6 @@ package mozilla.components.feature.session.bundling.db
 
 import mozilla.components.browser.session.Session
 import mozilla.components.browser.session.SessionManager
-import org.json.JSONObject
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertTrue
 import org.junit.Test
@@ -17,7 +16,7 @@ import org.robolectric.RobolectricTestRunner
 class BundleEntityTest {
     @Test
     fun `updateFrom updates state and time`() {
-        val bundle = BundleEntity(0, "", 0, UrlList(listOf()))
+        val bundle = BundleEntity(0, 0, UrlList(listOf()))
 
         val snapshot = SessionManager.Snapshot(
             listOf(SessionManager.Snapshot.Item(session = Session("https://www.mozilla.org"))),
@@ -29,13 +28,5 @@ class BundleEntityTest {
 
         assertEquals(1, bundle.urls.entries.size)
         assertEquals("https://www.mozilla.org", bundle.urls.entries[0])
-
-        val json = JSONObject(bundle.state)
-
-        assertEquals(1, json.get("version"))
-        assertEquals(0, json.get("selectedSessionIndex"))
-
-        val sessions = json.getJSONArray("sessionStateTuples")
-        assertEquals(1, sessions.length())
     }
 }